frontend improvements

+++ ContactsApp.Server/Controllers/ContactsController.cs
@@ -1,7 +1,10 @@
 ﻿using ContactsApp.Server.DTOs;
 using ContactsApp.Server.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace ContactsApp.Server.Controllers
 {
@@ -24,9 +27,11 @@ namespace ContactsApp.Server.Controllers
         /// <returns>
         /// Returns a detailed contact info for a contact with contactId.
         /// </returns>
+
         [HttpGet("{contactId}")]
         public async Task<ActionResult<ContactDto>> GetContactForUser(int contactId)
         {
+
             var contact = await _service.GetContactById(contactId);
 
             if (contact == null)
@@ -39,8 +44,18 @@ namespace ContactsApp.Server.Controllers
 
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<bool>> AddContact([FromBody] ContactDto contactDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (userId == null || int.Parse(userId) != contactDto.UserId)
+                return Unauthorized("Unauthorized Access");
+
+
             var result = await _service.AddContactAsync(contactDto);
 
             if (result == false)
@@ -49,9 +64,21 @@ namespace ContactsApp.Server.Controllers
             return Ok();
         }
 
+        [Authorize]
         [HttpPut("{contactId}")]
         public async Task<ActionResult<bool>> UpdateContact(int contactId, [FromBody] ContactDto contactDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            var contact = await _service.GetContactById(contactId);
+            var userIdFromContact = contact?.UserId;
+
+            if (userId == null || int.Parse(userId) != contactDto.UserId || int.Parse(userId) != userIdFromContact)
+                return Unauthorized("Unauthorized Access");
+
             var result = await _service.UpdateContactAsync(contactId, contactDto);
 
             if (result == false)
@@ -60,9 +87,22 @@ namespace ContactsApp.Server.Controllers
             return Ok();
         }
 
+        
+        [Authorize]
         [HttpDelete("{contactId}")]
         public async Task<ActionResult<bool>> DeleteContact(int contactId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            var contact = await _service.GetContactById(contactId);
+            var userIdFromContact = contact?.UserId;
+
+            if (userId == null || int.Parse(userId) != userIdFromContact)
+                return Unauthorized("Unauthorized Access");
+
             var result = await _service.DeleteContactAsync(contactId);
 
             if (result == false)
